feat: add client to query snapshot.debian.org
The snapshot.debian.org service provides detailed information about
packages, relations and versions. To enhance the SBOMs, we provide this
client implementation to query additional information (if needed).

The client internally uses the requests library to query the web API of
the snapshot. By that, we also add it to the required dependencies.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>