feat: add download command to retrieve packages from snapshot mirror
For license clearing it is important to get the source of a package
(both source and binary packages). For that, we add the subcommand
"download" to download all referenced debian packages to a local
directory. To avoid time consuming downloads on changes, we check
upfront if we already have the package and only download if not.

We further move the progress option to the generic options, as both
generate and download support it (and hopefully future commands will
support it as well).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>