feat(download): report cached data in .stat as well
On repeated executions some files already have been downloaded.
Currently these are already skipped when downloading again, but they
show up in the statistics. We now split the statistics into the total
data and the data we already have. This gives a better overview what
still is missing and needs to be downloaded. We further change the
return type to a named tuple to make it easier for downstream users to
access the various values.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>