feat(merger): add infrastructure to create combined debian archive
A source package consist of multiple individual parts (e.g. policy 4.x
diff files, .orig and .debian tarballs). To create a single artifact
which can be used for license clearing, we need to merge these into a
single archive. For that, we introduce the SourceArchiveMerger class
which performs the merge based on the .dsc data.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>