perf(download): use single http session
To resolve packages, we issue a lot of tiny requests to the snapshot
mirror. To speed this up, we now use a session to pool the http
connections and only resolve DNS once. We further re-use the same
session when downloading as this will help us to track rate limits in
the future.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>