refactor(dpkg): use a base package for binary/source packages
Introduce a `Package` class as a base for binary and source packages.
When parsing the dpkg status file now return all binary and source
packages, instead of only binary ones. This is a better representation
of the data, since in the SBOMs every source or binary package is an
entry in the SBOM.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>