refactor: move bom reference logic to a separate class
The logic for creating BOM references now lives in a separate Reference
class. This separates the concerns more cleanly.

Also do not keep copies of Dependencies and Sources in the binary
packages, but instead simply a Reference. As the matching of
relationships is only done with the BOM references anyways, this is
also enough for our internal data structures.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>