feat(generate): track built-using relations
Binary packages that contain data from other packages via a copy-in or
static link must be tracked as well, as otherwise the sources of the
copied-in data are now tracked at all. Debian provides the built-using
field for that, which contains a list of versioned source packages that
were used to generate the binary. We now track this relation as well and
add the corresponding source packages to the sbom.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>