refactor: track dependencies on packages instead of references
By tracking the precise dependencies of our dpkg.Package instances, we
can model cases where multiple versions of a package are referenced.
Further, we can clearly separate the parsing stages by not creating
referenced packages on-the-fly (follow up commit). This also allows to
decouple the spdx / cdx representation from the package list, as the
SBOM writers now can decide on their own way of issuing IDs to package
instances.

While refactoring this we noticed, that we currently do not precisely
track the source packages in their used versions (e.g. two libraries
share the same source package, but not in the same version). To only
refactor things, we keep this inconsistency by adding corresponding hash
and equality comparision operators to the SourcePackage class. Later on,
these can be made more precise, once we also track source package
versions in the dependencies.

This also requires us to refactor the unit test, as they test internals.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>