feat(generate): allow multiple versions of the same source package
As a preparation to track built-using relations, we need to allow the
listing of a single source package in multiple versions. This is because
the debian package mirror is not bootstrapped on new versions, but
instead incrementally updated. By that, a package and its corresponding
source package can be more recent that a package that was previously
built using copy-in sources from an older version of the same source
package.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>