feat(generate): add multi-arch support
Debian allows to install the same binary package for multiple architectures
at a time. As these packages are still different, they cannot be mapped
to the same entry in the SBOM. We reflect that by adding the
architecture to the package ID. When resolving cross binary package
dependencies, we further need to perform lookups to locate the exact
package another package depends on. This is performed based on the
debian rules (e.g. same architecture or arch=all).

As the reference ids changed, we also need to update the unit tests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>