chore: fix all type hints according to beartype
This is one major commit to do the whole cleanup, as this is hard to be
broken down. We further convert the deprecated typing module types to
the PEP 585 types, which are usable from Python 3.10 on. As our SBOM
dependency anyways is not compatible with older python versions on
Debian, this is a safe choice.

We further enable the beartype check of the debsbom module in our unit
tests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>