chore: make download dependencies optional
The core feature of the debsbom tool is the generate command. This is
usually run in build systems and by that we want to limit the set of
dependencies to a bare minimum for the sbom generation.

We implement that by making the dependencies needed for download and
source-merge optional. If they are not found at runtime, the
corresponding commands are disabled.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>