perf(generate): use filter function to limit set of apt packages
When parsing the apt data, we only need to create source packages for
packages we actually have in our list. Profiling has shown, that we
spend a lot of time in package creation. To improve this, we now pass a
filter function down to the apt parsing stage and only create packages
with names we have in our package list.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>