chore: explicitly check for python-apt to avoid internal warning
The dep822 library internally uses python-apt to significantly speedup
the parsing of Dep822 data. The API of dep822 allows to parse with an
alternative (but much slower parser), however this explicitly needs to
be requested as otherwise a warning is issued. For that, we need to
check if we have that dependency upfront, which we do here.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>