chore(generate): prepare package import for apt-cache binary packages
As a preparation to also extend our binary packages with data from the
apt cache, we rework the package import handling. As the number of
binary packages the apt cache knows is potentially huge, we further
rework the code to ensure we process the apt-cache data in O(1) space
and O(n) time. For that, we iterate the apt cache and lookup our own
packages instead of vice-versa.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>