chore(package): add filename and locator property
For binary packages, this refers to the name of the .deb file. For
source packages either to the .dsc file or to a explicitly set name.
By that, we can easily attach filename information to the components of
the sbom. If we further have a precise locator (e.g. a path on the
snapshot mirror or a local path), we further can attach it to the
package directly.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>