feat: add command to create a standard bom source archive
The standard-bom archive is a CDX standard reserved by Siemens and
provides a way to create a single artifact with an SBOM and all sources
and binaries. The repack command can be used to aggregate the sources,
copy / link them into the corresponding directory layout and recreate
the SBOM with the new references.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>