feat(generate): add source packages to CycloneDX format
CycloneDX v1.6 does not have a proposed way of expressing the relation
between a source package and a binary package in a machine readable way.
Until this is fixed in the spec, we add the source packages similar to
binary packages, where they can be distinguished by the PURL. We further
add a dependency from the binary to the source package.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>