chore(apt-cache): enrich package information with extended states
The apt extended states track if a package is manually installed (via
apt) or installed as a dependency. We now track this to later use that
information when creating the dependency graph in the SBOM.

Note, that this implementation uses a filter function to only track
states of packages we care about.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>