feat(generate): support taking a list of packages as input
We currently only allow to parse a dpkg status file. For that, the
packages need to be installed. There are situations where we have
the apt-cache and a list of packages and just want to create an sbom of
these.

We now add support to specify a list of binary packages with their
versions and architecture, perform the lookup in the apt cache and
create the SBOM.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>