fix(generate): avoid trailing space in supplier name
The suppliers are derived from the package maintainer, which is an
RFC 5322 string like "Jon Doe <john.doe@example.com>" or just a name.

With the currently used regex, the name contains the trailing whitespace
delimiting the name and the address. As it is hard to encode stripping
the trailing whitespace in Regex without lookahead expressions, we just
post-process the string and strip it.

Fixes: 958c97d ("Initial commit")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>