refactor(apt): rework compressed file handling
Rework how we handle maybe compressed files. Factor out the logic
dealing with maybe-compressed files. Use a different code path for
compressed and non-compressed files, so we do not need to use a `cat`
subprocess when reading non-compressed files.

This is also in preparation for a bug fix with to do with python-apt and
compressed files in later work.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>