chore(resolver): compress cache files with zstd
Tests on some systems show that the cache can get quite big (50MB or
more) if packages which have multiple references on the snapshot mirrors
are processed (e.g. firefox-esr). To reduce the impact, we compress the
cache with zstd. This library is anyways almonst on all systems and will
be part of the Python standard library from 3.14 on.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>