fix(apt): fix compression when python-apt is available
When python-apt is available, and thus apt_pkg, the iter_paragraphs
method only accepts filehandles in python-debian 0.1.49. Contrary to
what the docs state, passing a filehandle to an encrypted file causes
a segmentation fault. Thus we need to treat it differently.

When dealing with compressed files we now decompress the file and never
us python-apt.

For python debian >= 1.0.0 this behavior is fixed, and the method also
accepts filenames. Debian bookworm still uses 0.1.49, so we have to
have this workaround for now.

Fixes: 2d070dc feat(apt): allow reading of compressed apt-cache files
Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>