chore(ci): run test suite with python-apt enabled
Run our tests and the smoke tests with python-apt. As python-apt is not
available as a PyPI package and can not be installed with pip, rely on
the package provided by apt. This means we need to use the distribution
provided python, and have to skip the setup-python action.

Add a simple additional test step that checks if the package is actually
available in python.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>