fix(package): avoid duplicated dependencies
Packages can have versioned dependencies, whereby a dependency is listed
multiple times but with different version selectors (e.g. >> 2.41, <<
2.42). According to the internal representation in debsbom, this still
is a single dependency. However, we currently add it as two, leading to
parallel edges (references) in the SBOM.

We fix this by adding a dynamic property to the BinaryPackage that
returns the unique list of dependencies (according to what can be
installed in parallel on a system).

Fixes: 958c97d ("Initial commit")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>