feat(export): add module to export SBOM to graph formats
Exporting the SBOM to a precise and annotated graph enables
graph-tooling to further analyze and reason about it.

We add support to parse a debsbom generated SPDX and CycloneDX SBOM and
convert that to graphml. The infrastructure is prepared to be easily
extendable for other graph formats like GXL or dot.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>