chore(ci): rework when we run our workflows
Run workflows for pull requests and for pushes to main. We previously
only ran it for pushes, which works fine for contributors who have
direct repository access, but not when you contribute from a fork.

As a side effect we no longer run CI for non-main branches in the
repository.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>