fix(generate): consider version when looking through apt cache
When looking for packages in the apt cache the version is now considered
for matching purposes. That means if a package with the same name and
architecture but a different version is available in the apt cache, it
will now not be enriched with it.

Previously we matched solely on the name and architecture, which lead to
wrong package information when specifying packages from a packages list.

Fixes: 36b86c1 ("chore(generate): prepare package import for apt-...")
Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>