fix(merger): merge package metadata also if package is already merged
The merge command checks if we already have the merged artifact and does
not merge again. This allows us to cancel-and-continue the merging on a
per-package basis and in general speeds up re-invocations.

However, we must always merge the metadata as parts (like the supplier)
might be missing in the input SBOM and need to be loaded from the DSC
file.

To fix this, we move the already-merged short cirtcuit directly before
the actual tarball merging operation. By that, also the checksum
calculation always happens - which is a safety measure against
accidential corruptions.

Fixes: ae8f95a ("feat(repack): enhance packages with data from dsc ...")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>