fix: rework handling of package list packages and apt cache
Do not treat the merging with apt cache data differently when handling
packages from a package list. The only difference we need to consider is
that we need to add referenced source packages if we deal with a package
list stream. As an optimization, this additional step is only done in
this case.

Previously the separate handling of package list packages lead to
them not being included in the SBOM when they did not have a matching
entry in the apt cache. Also we never added extended state information.

Fixes: eaa6c76 ("feat(generate): support taking a list of packages...")
Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>