refactor(resolver): make snapshot resolver standalone class
Resolving packages on the debian snapshot mirrors is currently done via
a static method of the PackageResolver base class. However, as the
package resolvers are now much more generic ingress modules this does
not fit anymore. Further, it makes it impossible to decouple the ingress
module from the download parts.

We now move the .resolve method into its own class UpstreamResolver that
also does not share a base-class with PackageResolver. We further make
the interface more object-oriented by moving the generic initialization
into the constructor.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>