refactor(dpkg): accept arbitrary input streams for ingress
As a preparation to also accept dpkg-status file data via stdin, we add
a peekable wrapper around the input stream. With that, we will be able
(future commit) to dispatch to python apt if the first ingress line (or
couple of bytes) indicates that a dep822 format is following.

By that, we also need to refactor the tests of that internal API.
While the test data is pure text data and the ingress also works with
text data, for technical reasons StringIO cannot be used prior to Python
3.14 as it misses required members. As a workaround, we just use
BytesIO, which is also accepted by our ingress.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>