feat(dpkg): also allow to pass dpkg status file via stdin
In case a user only has a dpkg status file (but nothing else), he
currently needs to construct the directory structure of a chroot with
just that single file. By extending the stdin ingress module to also
allow dpkg-status file-like data, the user does not need to create a
chroot-like directory structure in this case. As this involves
intermediate streams and consumers might be interested in the source of
data (e.g. a dpkg-status file), we wrap the package iterators in a
class.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>