feat: add merge command
Add a new command that hierarchically merges any number of SBOMs. The
merge is intended to be used for combining multiple SBOMs that describe
a Debian-based Linux distribution. A common such use-case would be
combining SBOMs for a rootfs and a initrd.

The merging itself takes the root components/packages of the SBOMs and
moves them to the first dependency level. Any dependencies below that
are left untouched. That would mean something like this:

doc1-root
|- binary-dep1
|  |- source-dep1
|- binary-dep2

doc2-root
|- binary-dep3
|  |- source-dep3
|- binary-dep4

Would turn into this:

merged-doc-root
|- doc1-root
|  |- binary-dep1
|  |  |- source-dep1
|  |- binary-dep2
|- doc2-root
|  |- binary-dep3
|  |  |- source-dep3
|  |- binary-dep4

Duplicate components/packages are identified by their PURL. If such a
component/package is found then their SBOM reference ID is replaced in
the merged document. The identifiers are taken from the first occurrence
in the passed SBOMs. The component/package will also be deduplicated and
correctly referenced in all merged SBOMs.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>