feat(dpkg): track package installation status
Packages can be listed in the dpkg status file, despite not being
installed. This e.g. happens on system updates when config files of
removed packages are still present. Along that, there are a couple of
other not so common states.

We now parse and track these states per package in our internal
representation to let consumers (e.g. the SBOM serializers) use that
information.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>