feat: only add installed packages to sbom
Packages that are present in the status file but not installed on the
system do not need to be presented in the SBOM (or at least not as full
packages with a PURL). For now, we just completely skip them. Later on
this could be modeled more fine-grained by tracking which files are
still present.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>