docs(download): add JSON output schema
Users should not have to read the code or guess from the output what the
format looks like.
The .gitignore file was changed to only ignore JSON files in the root
directory since most of the temporary files are located there. We can't
just ignore all JSON files because we need them for the documentation.

Signed-off-by: Tamino Larisch <tamino.larisch@siemens.com>