refactor(dpkg): add helper to separate src and bin packages
We currently have many locations in our codebase where we explicitly
check for the concrete instance of package, just to see if it is a
source or binary package. To avoid this antipattern, we add helpers in
the package class to perform this check. We further often only want to
iterate the binary packages or the source packages. For that, we add
free functions to perform the filtering.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>