fix(generate): when using a package list do not use extended states
When specifying packages with a package list the extended states file is
not used to mark packages as (not) manually installed. Otherwise you
often run into the case where the generated SBOMs are missing the
connection between the root and the specified packages.

This is the correct behavior as packages from a package list are not
expected to be related to the system we use the apt cache from. So the
apt cache might still be useful to fill out missing package
information, but the extended states information is not relevant for
them.

Fixes: 71d02d0 ("feat(generate): track and use extended states ...")
Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>