feat(package): track checksums of dsc files for source packages
In Debian, the integrity of a source package is ensured by the integrity
of the dsc file, as this file references all artifacts. By that, we
consider this as the checksum of a source package. The apt, cache has
this data, so add it to our internal representation as well.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>