feat(repacker): check sha256sum of dsc file if present
In case we have this checksum in the SBOM, we compare it with the one of
the downloaded dsc file. By that, we also ensure the integrity of the
source packages. If not present, we skip this check.

As of now, we only compare the sha256sums across the codebase. Later on,
this check could be improved by comparing all checksums by order /
quality of the algorithm. This is not yet in scope.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>