feat(download): fetch artifacts for all archives
There are situations where a source package is uploaded to multiple
archives (e.g. debian and debian-ports), but the content differes even
within a version. This currently leads to incorrect checksums, as the
downloader only considers the first observed file.

We change this by downloading all artifacts and placing them in a
per-archive directory. This also requires slight changes to the
source-merge and repack, as now the download directory layout is
different.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>