chore(resolver): on imprecise lookup return most-recent files
When doing a lookup of a source package, whereby the source package has
multiple name / version identical entries on the snapshot mirror (e.g.
because it was re-signed over time without changing the content), we now
only return the most-recent version. By that, we avoid the name
collisions in the download directory. We further can't do any better, as
we do not know which version was used when generating the SBOM.

The case where we have precise data (e.g. a checksum) is not changed.
There, we still iterate all artifacts and return the one with matching
checksum.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>