fix(reference): correctly resolve dependencies of arch-all packages
Architecture all packages can depend on architecture native packages.
This was missed when adding the multi-arch support, as only the
direction arch=<...> -> arch=all was mapped but not the other way round.
As we might have multiple architecture specific dependencies with the
same name and version at a time, we need to explicitly resolve to the
native arch. We now read this from the rootfs if not explicitly
specified by the caller.

Fixes: 38ddc7e ("feat(generate): add multi-arch support")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>