chore(download): change imprecise lookup warning level to info
The imprecise lookup warning happens when no checksums are available for
a package. In theory the package should be uniquely identifiable by just
their name and version (and arch, if binary), without the use of
checksums. Everything else should be a bug in the snapshot mirror. In
this case downgrading from a warning to an info message seems
appropriate, since it is an upstream bug if it happens and we would
emit the warning almost always otherwise.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>