fix(generate): fix residual config packages being placed into SBOMs
Packages that are not properly installed but have some residual config
still show up in the dpkg status file. We do not want to place these
packages into the SBOM. As it turns out we never properly removed them
but instead only disabled dependency resolution.

Since any non-installed binary packages might pull in source packages we
have to remove them early on in the parsing process. This means that
only installed packages are now passed to the generator functions.

Fixes: 0325401 ("feat: only add installed packages to sbom")
Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>