chore(tests): register download files with package relation
Currently, a download item does not have to be linked to a package.
While it is always linked to a package throughout the codebase, the
tests take shortcuts and do not always provide a related package.

As a preparation to make the package mandatory when registering download
files, we adapt the tests to always pass a package.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>