feat: add Vcs information to sbom
For cdx add the Vcs information and the url to the web interface to the
'external_refs' object with the type set to 'vcs'.

For spdx there there isn't a suitable category or type for this in the
externalReferences field. Therefore, we use the Other category with the
custom types 'vcs' and 'vcsBrowser'.

This Vcs information can be useful for analysing the packages (e.g. how
well a package is maintained and what the risk factor is).

Signed-off-by: Tamino Larisch <tamino.larisch@siemens.com>