refactor(download): generalize upstream resolver logic
As a preparation for upcoming work untangle the logic surrounding the
upstream resolver. For this create a generic base class for resolving
and a generic remote file representation. Move logic related to the
snapshot mirror to the snapshot client module.

Also factor out the resolve logic itself: there is a base part which
handles caching of packages and one specific to the actual resolver.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>