refactor(download): only use required fields from the remote file
With the refactored remote file structure it is not possible to derive
the target path directly from the remote file. For this information
about the package, namely if it is source or binary, is required.

The size is also optional now, so the calculation of download sizes
needs to consider that too. For files where the size is unknown the size
is 0.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>