refactor(download): make cache replaceable
Allow a cache implementation to be replaced while in use. This is
required so we do not have to pass the default cache instance to the
initialization.

This is in preparation for the upcoming plugin work. The use-case is to
allow for custom cache implementations and, in case no custom
implementation is provided, use our internal default implementation.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>