refactor: standardize RemoteFile checksum storage
RemoteFile now uses our intern mapping of ChecksumAlgo to str, as it is
in most places where we have checksums. Apart from removing the
ambiguitiy of what hash algorithm 'hash' uses, it also allows the use of
the new checksum verification methods to compare checkusms.
This change requires the use of a frozenset in PackageDownloader to
ensure that files are identified by their complete set of checksums.

Signed-off-by: Tamino Larisch <tamino.larisch@siemens.com>