feat: make cdx and spdx dependencies optional
Previously, both libraries were always required, even if a user only
intended to use one format, which would lead to errors if the unused
dependency was missing. Now, users only need to install the specific
dependency for the SBOM format they wish to use. This also means that
missing dependency errors are now more precise, occurring only when a
user tries to execute a command for a format whose required library is
not installed.
This update may simplify packaging efforts for
distributions like Debian, as `debsbom` can now be packaged with just
`python3-cyclonedx-lib` even if a `spdx-tools` package is not yet
available.

Signed-off-by: Tamino Larisch <tamino.larisch@siemens.com>