feat: add centralized check for API status codes
Introduces a new `get` helper function to centralize HTTP requests to
the snapshot client. This helper now validates the response's HTTP
status code. A 404 error code will raise a `NotFoundOnSnapshotError`.
All other client or server error codes will be caught by the
`raise_for_status` function from the `requests` library and subsequently
re-raised as a `SnapshotDataLakeError`. The only remaining status codes
that the snapshot client is expected to return are 200 and 304, where
the latter one cannot be returned, as an `If-Modified-Since` header is
never specified. This implies that if no error is raised, the status
code is 200 and the output is in the expected format.

This prevents silent failures and subsequent issues like incorrect JSON
parsing when the API returns unexpected content, ensuring earlier and
clearer error detection. It also reduces the need for repetitive status
code checks at each call site.

Signed-off-by: Tamino Larisch <tamino.larisch@siemens.com>