refactor(schema): provide download schema as variable on module import
Make `schema` a proper Python module and load `schema-download.json`
when the module is imported. This means that the schema can be
accessed as a "quasi-constant" via `debsbom.schema.download` (similar
to `math.pi`, for example).

Note that module-level code is only executed once, namely during the
initial import of the module, which means that the file is only loaded
into memory once.

Signed-off-by: Max R. Carrara <max@aequito.sh>