test(download): use download schema through module
Instead of loading the download result schema as fixture, access it
through the `schema` submodule of `debsbom` directly.

This avoids any unexpected errors when running the tests in an
environment where the directory structure may be slightly different,
for example when building and testing the `debsbom` package using
Debian's Python packaging tooling (pybuild specifically).

Signed-off-by: Max R. Carrara <max@aequito.sh>