build: add MANIFEST.in
Having a MANIFEST.in file is necessary so that setuptools can figure
out what files to include in the final sdist (source distribution).
By extension, this also lets Debian tooling like  pybuild / dh_python3
[0] infer what to include in the final binary Debian package.

The file contains a list of commands, each telling setuptools what
files to include or exclude in the sdist. [1]

In particular, this file ensures that
`src/debsbom/schema/schema-download.json` is included in the final
sdist, which is necessary for the `debsbom.schema` module.

[0] https://packages.debian.org/dh-python
[1] https://setuptools.pypa.io/en/latest/userguide/miscellaneous.html

Signed-off-by: Max R. Carrara <max@aequito.sh>