build(debian): add debian/control
Add the initial debian/control file necessary for packaging debsbom
for Debian.

The source and binary package names are chosen based on how a handful
of other Python tools are packaged [0][1][2], that is, "debsbom" for
both.

It should be noted that the X-Python3-Version field for the source
package stanza is added, as suggested by the Debian Python packaging
policy [3].

[0]: https://salsa.debian.org/python-team/packages/black/-/blob/592940bceecc090ff019c7e059a43f7590762442/debian/control
[1]: https://salsa.debian.org/python-team/packages/ruff/-/blob/3754df1dd55b476f470a88d24be5fe9a40177c4b/debian/control
[2]: https://salsa.debian.org/python-team/packages/mypy/-/blob/300f8d49ab58945999a93556364102662f0fc0cc/debian/control
[3]: https://www.debian.org/doc/packaging-manuals/python-policy/index.html#specifying-supported-versions

Signed-off-by: Max R. Carrara <max@aequito.sh>