build(debian): add debian/copyright
Add debian/copyright, which was initially generated using `decopy` [0]
and slightly modified to ensure correctness.

For reference, the complete command used to yield the initial
debian/copyright file is as follows:

    decopy --exclude '^(build|\.venv|\.mypy_cache|\.pytest_cache|\.github|dist)/*'

This command should be used to keep the debian/copyright file updated
as the repository grows. Any directories not actually part of the
source tree should be added to the regex of the `--exclude` flag in
order to ensure correctness of the copyright file. In particular, it
avoids inclusion of copyright holders that have not actually
contributed to the repository, but are instead found in dependencies
or similar (e.g. inside a virtual environment in `.venv/`).

Signed-off-by: Max R. Carrara <max@aequito.sh>