fix(generate): remove spaces in vcs locator fields
The SPDX specification mandates that the `locator` field within external
references must not contain spaces. This commit removes them and the
additional details (e.g., branch, path) that would follow a space.
Such auxiliary information is rarely present in packages and is
typically not critical for an SBOM. Therefore, these details are also
removed for CDX generation to ensure consistency.

fixes: 05ca235 ("feat: add Vcs information to sbom")

Signed-off-by: Tamino Larisch <tamino.larisch@siemens.com>