feat(generate): add license information to SBOMs
Licenses are converted from the Debian license information to SPDX
expressions. First the expression syntax is converted and afterwards
license symbols are replaced. This replacement step uses a lookup table
to prevent any errors in the conversion because it is important we
never get the license information wrong. The lookup table was based on
the syntax specification that Debian provides [1].

For SPDX we use the `licenseDeclared` field. This makes it clear we did
not perform any analysis but instead just repeated what the package
author stated.

For CDX we use the `evidence` field. We can similarly portray that the
license information comes from the original package author.

[1] https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/#license-specification

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>